Moved /send-magic-link rate limit tests to correct file
no issue

- `members/signin.test.js` is testing the `?token=` signin path but also contained the `/send-magic-link/` API endpoint rate-limit paths
- moved the rate limit tests to `members/send-magic-link.test.js` so there's not cross-purpose tests across both
