+++ ghost/core/test/e2e-api/members/send-magic-link.test.js
@@ -1,5 +1,6 @@
-const {agentProvider, mockManager, fixtureManager, matchers, configUtils} = require('../../utils/e2e-framework');
+const {agentProvider, mockManager, fixtureManager, matchers, configUtils, resetRateLimits, dbUtils} = require('../../utils/e2e-framework');
 const should = require('should');
+const sinon = require('sinon');
 const settingsCache = require('../../../core/shared/settings-cache');
 const settingsService = require('../../../core/server/services/settings');
 const DomainEvents = require('@tryghost/domain-events');
@@ -500,6 +501,243 @@ describe('sendMagicLink', function () {
             });
 
             should.exist(mail);
+        });
+    });
+
+    describe('Rate limiting', function () {
+        let clock;
+
+        beforeEach(async function () {
+            await dbUtils.truncate('brute');
+            await resetRateLimits();
+            clock = sinon.useFakeTimers(new Date());
+        });
+
+        afterEach(function () {
+            clock.restore();
+        });
+
+        it('Will rate limit member enumeration', async function () {
+            // +1 because this is a retry count, so we have one request + the retries, then blocked
+            const userLoginRateLimit = configUtils.config.get('spam').member_login.freeRetries + 1;
+
+            for (let i = 0; i < userLoginRateLimit; i++) {
+                await membersAgent.post('/api/send-magic-link')
+                    .body({
+                        email: 'rate-limiting-test-' + i + '@test.com',
+                        emailType: 'signup'
+                    })
+                    .expectStatus(201);
+            }
+
+            // Now we've been rate limited for every email
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'other@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Now we've been rate limited
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'one@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Get one of the magic link emails
+            const mail = mockManager.assert.sentEmail({
+                to: 'rate-limiting-test-0@test.com',
+                subject: /Complete your sign up to Ghost!/
+            });
+
+            // Get link from email
+            const [url] = mail.text.match(/https?:\/\/[^\s]+/);
+
+            const magicLink = new URL(url);
+
+            // Login works, but we're still rate limited (otherwise this would be an easy escape to allow user enumeration)
+            await membersAgent.get(magicLink.pathname + magicLink.search);
+
+            // We are still rate limited
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Wait 10 minutes and check if we are still rate limited
+            clock.tick(10 * 60 * 1000);
+
+            // We should be able to send a new email
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(201);
+
+            // But only once
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any2@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Waiting 10 minutes is still enough (fibonacci)
+            clock.tick(10 * 60 * 1000);
+
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any2@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(201);
+
+            // Blocked again
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any3@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Waiting 10 minutes is not enough any longer
+            clock.tick(10 * 60 * 1000);
+
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any3@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Waiting 20 minutes is enough
+            clock.tick(10 * 60 * 1000);
+
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any2@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(201);
+
+            // Blocked again
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any3@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Waiting 12 hours is enough to reset it completely
+            clock.tick(12 * 60 * 60 * 1000 + 1000);
+
+            // We can try multiple times again
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any4@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(201);
+
+            // Blocked again
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'any5@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(201);
+        });
+
+        it('Will clear rate limits for members auth', async function () {
+            // Temporary increase the member_login rate limits to a higher number
+            // because other wise we would hit user enumeration rate limits (this won't get reset after a succeeded login)
+            // We need to do this here otherwise the middlewares are not setup correctly
+            configUtils.set('spam:member_login:freeRetries', 40);
+
+            // We need to reset spam instances to apply the configuration change
+            await resetRateLimits();
+
+            // +1 because this is a retry count, so we have one request + the retries, then blocked
+            const userLoginRateLimit = configUtils.config.get('spam').user_login.freeRetries + 1;
+
+            for (let i = 0; i < userLoginRateLimit; i++) {
+                await membersAgent.post('/api/send-magic-link')
+                    .body({
+                        email: 'rate-limiting-test-1@test.com',
+                        emailType: 'signup'
+                    })
+                    .expectStatus(201);
+
+                await membersAgent.post('/api/send-magic-link')
+                    .body({
+                        email: 'rate-limiting-test-2@test.com',
+                        emailType: 'signup'
+                    })
+                    .expectStatus(201);
+            }
+
+            // Now we've been rate limited
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'rate-limiting-test-1@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Now we've been rate limited
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'rate-limiting-test-2@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Get one of the magic link emails
+            const mail = mockManager.assert.sentEmail({
+                to: 'rate-limiting-test-1@test.com',
+                subject: /Complete your sign up to Ghost!/
+            });
+
+            // Get link from email
+            const [url] = mail.text.match(/https?:\/\/[^\s]+/);
+
+            const magicLink = new URL(url);
+
+            // Login
+            await membersAgent.get(magicLink.pathname + magicLink.search);
+
+            // The first member has been un ratelimited
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'rate-limiting-test-1@test.com',
+                    emailType: 'signup'
+                })
+                .expectEmptyBody()
+                .expectStatus(201);
+
+            // The second is still rate limited
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'rate-limiting-test-2@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(429);
+
+            // Wait 10 minutes and check if we are still rate limited
+            clock.tick(10 * 60 * 1000);
+
+            // We should be able to send a new email
+            await membersAgent.post('/api/send-magic-link')
+                .body({
+                    email: 'rate-limiting-test-2@test.com',
+                    emailType: 'signup'
+                })
+                .expectStatus(201);
         });
     });
 });
